requirements.txt
Flask==1.1.2
openai_whisper==20231117
python-dotenv==1.0.1
Requests==2.32.3
